requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
binance==0.3.43
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.2
daphne==4.2.0
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_daq==0.6.0
dateparser==1.2.1
Django==5.2.1
django-plotly-dash==2.4.6
django-redis==5.4.0
dpd_components==0.1.0
Flask==3.0.3
frozenlist==1.6.0
h11==0.16.0
htmltools==0.6.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.4.4
narwhals==1.40.0
nest-asyncio==1.6.0
numpy==2.2.6
orjson==3.10.18
packaging==25.0
pandas==2.2.3
plotly==6.1.0
prompt_toolkit==3.0.51
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pyOpenSSL==25.1.0
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
questionary==2.1.0
redis==6.1.0
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
service-identity==24.2.0
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.0.6
whitenoise==6.9.0
yarl==1.20.0
zipp==3.21.0
zope.interface==7.2
